Workaround for APT ending up in a broken state.
Znake mounts the hosts /etc/passwd file into the container.
The Ubuntu Docker images makes assumptions about what users
are available. This causes APT to end up in a broken state
when a user is present in the statoverride file but is not
present in the /etc/passwd file that is mounted.

As a short term solution, empty the statoverride file.